Correctly handle panics during logging
Previously when a panic occurred during the formatting/logging of a
record, and with the log-panic feature enabled, it would log again. This
meant that the log function would be called again, but this time the BUF
thread-local was in use. Trying to borrow this thread-local again would
panic as shown in issue 42. The following code was enough to trigger the
bug.

    struct T;

    impl fmt::Display for T {
        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
            panic!()
        }
    }

    info!("T: {}", T);

This commit fixes the problem by allocating a new buffer in the rare
case this happens.